requirements.txt
PyOpenGL 
PyOpenGL_accelerate
numpy
stable-baselines3[extra]
matplotlib
torch
opencv-python
shimmy>=2.0